05 DAY JAVA02

+++ JAVA01/src/Ch02/C02TypeChange.java
@@ -3,13 +3,13 @@ package Ch02;
 public class C02TypeChange {
 
 	public static void main(String[] args) {
-		// 강제 형변환
-		// 좁은 범위 공간에 큰값을 넣으려고 하는 경우
-		// 기본적으로 불가능하기 때문에 강제로 자료형을 바꾸어 전달한다.
-		// 데이터 손실의 염려가 있다.
+		//강제 형변환
+		//좁은 범위 공간에 큰값을 넣으려고 하는경우
+		//기본적으로 불가능하기 때문에 강제로 자료형을 바꾸어 전달한다
+		//데이터 손실의 염려가 있다
 		int intValue = 20000;
 		char charValue = (char)intValue;
-		System.out.println(charValue);
+		System.out.println((int)charValue);
 		
 		long longValue = 500;
 		intValue = (int)longValue;
@@ -17,8 +17,8 @@ public class C02TypeChange {
 		
 		//데이터 손실!
 		double doubleValue = 3.14;
-		intValue = (int)doubleValue;
-		System.out.println(intValue);
+		intValue =  (int)doubleValue;
+		System.out.println(intValue);	
 	}
 
 }
